% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{cdf_bernoulli}
\alias{cdf_bernoulli}
\title{CDF Bernoulli}
\usage{
cdf_bernoulli(x, p)
}
\arguments{
\item{x}{Nilai x (real)}

\item{p}{Probabilitas sukses}
}
\value{
Probabilitas kumulatif hingga x
}
\description{
CDF Bernoulli
}
