% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{cdf_binomial}
\alias{cdf_binomial}
\title{CDF Binomial (dengan penjumlahan manual PMF)}
\usage{
cdf_binomial(x, size, p)
}
\arguments{
\item{x}{Nilai x}

\item{size}{Jumlah percobaan (n)}

\item{p}{Probabilitas sukses}
}
\value{
Probabilitas kumulatif hingga x
}
\description{
CDF Binomial (dengan penjumlahan manual PMF)
}
