% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{estimasi_p_bernoulli}
\alias{estimasi_p_bernoulli}
\title{Estimasi Parameter p untuk Distribusi Bernoulli}
\usage{
estimasi_p_bernoulli(data)
}
\arguments{
\item{data}{Vektor data bernoulli (0 atau 1)}
}
\value{
Estimasi p̂ sebagai rata-rata dari data
}
\description{
Estimasi Parameter p untuk Distribusi Bernoulli
}
