% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{estimasi_p_binomial}
\alias{estimasi_p_binomial}
\title{Estimasi Parameter p untuk Distribusi Binomial}
\usage{
estimasi_p_binomial(data, size)
}
\arguments{
\item{data}{Vektor jumlah sukses per observasi}

\item{size}{Jumlah percobaan per observasi}
}
\value{
Estimasi p̂ sebagai proporsi keberhasilan
}
\description{
Estimasi Parameter p untuk Distribusi Binomial
}
