% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{plot_bernoulli}
\alias{plot_bernoulli}
\title{Plot Distribusi Frekuensi Data Bernoulli}
\usage{
plot_bernoulli(data)
}
\arguments{
\item{data}{Vektor data Bernoulli (berisi 0 dan 1)}
}
\value{
Barplot frekuensi 0 dan 1
}
\description{
Plot Distribusi Frekuensi Data Bernoulli
}
