% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{plot_binomial}
\alias{plot_binomial}
\title{Analisis dan Visualisasi Data Binomial}
\usage{
plot_binomial(data, size)
}
\arguments{
\item{data}{Vektor jumlah sukses per pelanggan (misalnya dari 10 pelanggan)}

\item{size}{Jumlah percobaan per pelanggan (harus diketahui)}
}
\value{
Estimasi p̂ dan barplot jumlah sukses
}
\description{
Analisis dan Visualisasi Data Binomial
}
