% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{plot_cdf_bernoulli}
\alias{plot_cdf_bernoulli}
\title{Plot CDF Distribusi Bernoulli}
\usage{
plot_cdf_bernoulli(p)
}
\arguments{
\item{p}{Probabilitas sukses}
}
\description{
Plot CDF Distribusi Bernoulli
}
