% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{plot_cdf_binomial}
\alias{plot_cdf_binomial}
\title{Plot CDF Distribusi Binomial}
\usage{
plot_cdf_binomial(size, p)
}
\arguments{
\item{size}{Jumlah percobaan}

\item{p}{Probabilitas sukses}
}
\description{
Plot CDF Distribusi Binomial
}
