% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{plot_pmf_bernoulli}
\alias{plot_pmf_bernoulli}
\title{Plot PMF Distribusi Bernoulli}
\usage{
plot_pmf_bernoulli(p)
}
\arguments{
\item{p}{Probabilitas sukses}
}
\description{
Plot PMF Distribusi Bernoulli
}
