% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{plot_pmf_binomial}
\alias{plot_pmf_binomial}
\title{Plot PMF Distribusi Binomial}
\usage{
plot_pmf_binomial(size, p)
}
\arguments{
\item{size}{Jumlah percobaan}

\item{p}{Probabilitas sukses}
}
\description{
Plot PMF Distribusi Binomial
}
