% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{pmf_bernoulli}
\alias{pmf_bernoulli}
\title{PMF Bernoulli}
\usage{
pmf_bernoulli(x, p)
}
\arguments{
\item{x}{Nilai 0 atau 1}

\item{p}{Probabilitas sukses}
}
\value{
Probabilitas f(x)
}
\description{
PMF Bernoulli
}
