% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{pmf_binomial}
\alias{pmf_binomial}
\title{PMF Binomial (dengan rumus eksplisit)}
\usage{
pmf_binomial(x, size, p)
}
\arguments{
\item{x}{Nilai x (jumlah sukses)}

\item{size}{Jumlah percobaan (n)}

\item{p}{Probabilitas sukses}
}
\value{
Probabilitas f(x)
}
\description{
PMF Binomial (dengan rumus eksplisit)
}
