% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{simulasi_bernoulli}
\alias{simulasi_bernoulli}
\title{Simulasi Bernoulli}
\usage{
simulasi_bernoulli(n, p, plot = FALSE)
}
\arguments{
\item{n}{Jumlah observasi}

\item{p}{Probabilitas sukses}

\item{plot}{Logika, apakah hasil akan diplot?}
}
\value{
Vektor data Bernoulli
}
\description{
Simulasi Bernoulli
}
