% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{simulasi_binomial}
\alias{simulasi_binomial}
\title{Simulasi Binomial}
\usage{
simulasi_binomial(n, size, p, plot = FALSE)
}
\arguments{
\item{n}{Jumlah observasi}

\item{size}{Jumlah percobaan}

\item{p}{Probabilitas sukses}

\item{plot}{Logika, apakah hasil akan diplot?}
}
\value{
Vektor data Binomial
}
\description{
Simulasi Binomial
}
